chore: reorder using and format files

+++ Ark.Tools.Core/Ex.cs
@@ -19,9 +19,9 @@ namespace Ark.Tools.Core
             {
                 string s = i.ToString("X2", CultureInfo.InvariantCulture);
                 if (BitConverter.IsLittleEndian)
-                    result[i] = ((uint)s[0]) + ((uint)s[1] << 16);
+                    result[i] = s[0] + ((uint)s[1] << 16);
                 else
-                    result[i] = ((uint)s[1]) + ((uint)s[0] << 16);
+                    result[i] = s[1] + ((uint)s[0] << 16);
             }
             return result;
         }
